requirements.txt
python-jenkins
boto3
